replication: moved replication.pb.go to replication folder
as we are planning to add more protobuf files that
are not related to each other like, replication,
Container orchestrator specific and fencing etc.
moving the replication to a new folder called
replication to keep the package name and directory
name similar.

Command to run to generate *.pb.go with the interface.

```bash
$ protoc --go_out=plugins=grpc:lib/go/replication *.proto
```
Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>