ci: add mergify configuration
This configuration adds a `design` label to any change that updates
files under `lib/go/`. Changes to designs are required to re-generate
the API, and should trigger adding the `design` label.

When the `deisgn` label is set, at least two approvals are required. For
other changes, a single approval is sufficient.

Signed-off-by: Niels de Vos <ndevos@redhat.com>