replication: change replication_secret to csi.v1.csi_secret
This commit changes replication secret to csi.v1.csi_secret
similar to network fence, to keep it consistent across layers.

Signed-off-by: Rakshith R <rar@redhat.com>

replication: add rpc to get volumereplication info

This commit adds new rpc to get the volume
repliction info which at this moment, is mainly to
get the last_sync_time. we can update this later
as per the requirement.

Signed-off-by: yati1998 <ypadia@redhat.com>

replication: define error conditions for GetVolumeReplicationInfo

In some cases the driver might not be able to return the
Replication details for a period of time, in that case
it can return well known error codes based on that CO
can retry to call the GetVolumeReplicationInfo in exponential
backoff to get the required details.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>

Add replication source field

Adjustments after comments on PR

Signed-off-by: Elena Gershkovich <elenage@il.ibm.com>