fence: add RPC to get the client address
Adding a RPC to get the local client details
which includes the id and the local ip addresses
for fencing. The ID can be anything that can be
used as an identifier about the client address.
For example the ID can be the cephcluster UUID
and the ip addressed/cidr is the local client
addresses used to connect to the ceph cluster,
The ceph is an example/representation here.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>